pyproject.toml
[tool.poetry]
name = "pymoysklad"
version = "0.1.2"
description = "very-WIP wrapper for Moysklad JSON API"
authors = ["Anatoly Raev <cralixraev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
requests-ratelimiter = "^0.3.2"
mashumaro = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.971"
python-dotenv = "^0.20.0"
types-requests = "^2.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

